+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `Atata.Cli.Npm` package v1.1.0
+- Add `Atata.Cli.Npm` package v1.1.0.
+
+### Changed
+
+- Inherit `HtmlValidateCli` from `GlobalNpmPackageCli<HtmlValidateCli>` instead of `ProgramCli<HtmlValidateCli>`.
 
 ## [1.0.0] - 2021-06-25